move the 'meta logic' of WF out to the db type
