+++ src/error.rs
@@ -1,3 +1,4 @@
+use crate::rules::wf::WfError;
 use failure::Error;
 
 /// Wrapper type for the various errors that can occur during chalk
@@ -11,6 +12,14 @@ pub struct ChalkError {
 
 impl From<Error> for ChalkError {
     fn from(value: Error) -> Self {
+        ChalkError {
+            error_text: value.to_string(),
+        }
+    }
+}
+
+impl From<WfError> for ChalkError {
+    fn from(value: WfError) -> Self {
         ChalkError {
             error_text: value.to_string(),
         }
